requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
langgraph
langsmith
langchain_anthropic
python-dotenv
langchain-openai
tavily-python
langchain_community
bs4
jupyter
tqdm
